Added assembly and processing

+++ Language/scr/warnings.h
@@ -1,13 +1,13 @@
 #ifndef WARNINGS_H
 #define WARNINGS_H
 
-#define INFORMATION_ABOUT_CALL (call_of_dump){__FILE__, (size_t)__LINE__, __FUNCTION__}
+#define INFORMATION_ABOUT_CALL (call_of_dump){__FILE__, __LINE__, __FUNCTION__}
 
 const size_t SIZE_OF_WARNINGS = 128;
 
 struct call_of_dump {
     const char* name_file;
-    size_t number_of_line;
+    int number_of_line;
     const char* name_function;
 };
